# docs_backend/openapi.json
{"openapi":"3.1.0","info":{"title":"RAG System API","description":"API para sistema de Retrieval Augmented Generation","version":"0.1.0"},"paths":{"/query":{"post":{"summary":"Query Documents","description":"Realiza uma consulta aos documentos usando RAG.","operationId":"query_documents_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/text":{"post":{"summary":"Add Text Document","description":"Adiciona um documento de texto à base de conhecimento.","operationId":"add_text_document_documents_text_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TextDocument"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/directory":{"post":{"summary":"Load Directory","description":"Carrega documentos de um diretório com logs detalhados.","operationId":"load_directory_documents_directory_post","parameters":[{"name":"directory","in":"query","required":true,"schema":{"type":"string","title":"Directory"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/image":{"post":{"summary":"Upload Image","description":"Faz upload e processa uma imagem.","operationId":"upload_image_documents_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_image_documents_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/images/directory":{"post":{"summary":"Load Image Directory","description":"Carrega e processa imagens de um diretório.","operationId":"load_image_directory_documents_images_directory_post","parameters":[{"name":"directory","in":"query","required":true,"schema":{"type":"string","title":"Directory"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stats/basic":{"get":{"summary":"Get Basic Stats","description":"Retorna estatísticas básicas da coleção com logs detalhados.","operationId":"get_basic_stats_stats_basic_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionStats"}}}}}}},"/stats/detailed":{"get":{"summary":"Get Detailed Stats","description":"Retorna estatísticas detalhadas da coleção com logs abrangentes.","operationId":"get_detailed_stats_stats_detailed_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetailedStats"}}}}}}},"/documents":{"get":{"summary":"Get Documents","description":"Lista todos os documentos ou documentos de uma fonte específica.","operationId":"get_documents_documents_get","parameters":[{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentContent"},"title":"Response Get Documents Documents Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Clear Collection","description":"Limpa toda a coleção com logs detalhados e rastreamento.","operationId":"clear_collection_documents_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/documents/export":{"post":{"summary":"Export Collection","description":"Exporta a coleção para um diretório.","operationId":"export_collection_documents_export_post","parameters":[{"name":"directory","in":"query","required":true,"schema":{"type":"string","title":"Directory"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/documents/{source}":{"delete":{"summary":"Remove Document","description":"Remove um documento específico da coleção.","operationId":"remove_document_documents__source__delete","parameters":[{"name":"source","in":"path","required":true,"schema":{"type":"string","title":"Source"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images/batch":{"post":{"summary":"Process Image Batch","description":"Processa um lote de imagens em paralelo com cache.","operationId":"process_image_batch_images_batch_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_process_image_batch_images_batch_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images/cache":{"delete":{"summary":"Clear Image Cache","description":"Limpa o cache de processamento de imagens.","operationId":"clear_image_cache_images_cache_delete","parameters":[{"name":"older_than","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Older Than"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/message":{"post":{"summary":"Chat Message","description":"Processa uma mensagem do chat usando RAG + LLM","operationId":"chat_message_chat_message_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessage"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat/history":{"get":{"summary":"Get Chat History","description":"Retorna o histórico do chat","operationId":"get_chat_history_chat_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/chat/clear":{"post":{"summary":"Clear Chat History","description":"Limpa o histórico do chat","operationId":"clear_chat_history_chat_clear_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_process_image_batch_images_batch_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_process_image_batch_images_batch_post"},"Body_upload_image_documents_image_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_image_documents_image_post"},"ChatMessage":{"properties":{"content":{"type":"string","title":"Content"},"include_context":{"type":"boolean","title":"Include Context","default":true}},"type":"object","required":["content"],"title":"ChatMessage","description":"Modelo para mensagem do chat"},"ChatResponse":{"properties":{"message":{"type":"string","title":"Message"},"success":{"type":"boolean","title":"Success"}},"type":"object","required":["message","success"],"title":"ChatResponse","description":"Modelo para resposta do chat"},"CollectionStats":{"properties":{"total_documents":{"type":"integer","title":"Total Documents"},"document_types":{"additionalProperties":{"type":"integer"},"type":"object","title":"Document Types"}},"type":"object","required":["total_documents","document_types"],"title":"CollectionStats"},"DetailedStats":{"properties":{"total_documents":{"type":"integer","title":"Total Documents"},"document_types":{"additionalProperties":{"type":"integer"},"type":"object","title":"Document Types"},"sources":{"items":{"type":"string"},"type":"array","title":"Sources"},"last_modified":{"type":"string","title":"Last Modified"},"collection_name":{"type":"string","title":"Collection Name"}},"type":"object","required":["total_documents","document_types","sources","last_modified","collection_name"],"title":"DetailedStats"},"DocumentContent":{"properties":{"source":{"type":"string","title":"Source"},"type":{"type":"string","title":"Type"},"content":{"type":"string","title":"Content"}},"type":"object","required":["source","type","content"],"title":"DocumentContent"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QueryRequest":{"properties":{"question":{"type":"string","title":"Question"}},"type":"object","required":["question"],"title":"QueryRequest"},"QueryResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"sources":{"items":{"type":"object"},"type":"array","title":"Sources"}},"type":"object","required":["answer","sources"],"title":"QueryResponse"},"TextDocument":{"properties":{"content":{"type":"string","title":"Content"},"metadata":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","required":["content"],"title":"TextDocument"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}